multi: move lnd client setup and feature checks into lnd module
